# specs/001-md/contracts/api.yaml
openapi: 3.0.3
info:
  title: MOF数据t-SNE交互式可视化 API
  description: 材料科学研究用MOF数据t-SNE降维和可视化Web应用API
  version: 1.0.0
  contact:
    name: MOF Visualization Team
    email: team@mof-viz.com
servers:
  - url: http://localhost:8000
    description: 本地开发服务器
paths:
...
  /api/upload:
    post:
      summary: 上传CSV文件并开始处理
      description: 用户上传包含MOF数据的CSV文件，系统自动识别格式并开始预处理
      operationId: upload_csv_file
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
                  description: CSV文件，最大100MB
                encoding:
                  auto_detect:
                    contentType: text/csv
      responses:
        '200':
          description: 文件上传成功，返回数据集信息
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '413':
          $ref: '#/components/responses/FileTooLarge'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
...
  /api/datasets/{dataset_id}/process:
    post:
      summary: 启动数据处理流水线
      description: 启动t-SNE降维处理流程，支持参数配置
      operationId: start_processing
      parameters:
        - name: dataset_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProcessingRequest'
      responses:
        '200':
          description: 处理任务已启动
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProcessingResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: 数据集正在处理中
...
components:
  schemas:
...
    BaseResponse:
      type: object
      properties:
        success:
          type: boolean
          description: 请求是否成功
        timestamp:
          type: string
          format: date-time
          description: 响应时间
        request_id:
          type: string
          format: uuid
          description: 请求唯一标识符
    ErrorResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            error:
              type: object
              properties:
                code:
                  type: string
                  description: 错误代码
                message:
                  type: string
                  description: 错误信息
                details:
                  type: object
                  description: 错误详情

    # 数据相关Schema
    UploadResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                dataset_id:
                  type: string
                  format: uuid
                filename:
                  type: string
                total_rows:
                  type: integer
                total_columns:
                  type: integer
                encoding:
                  type: string
                data_quality_score:
                  type: number
                  format: float
                  minimum: 0
                  maximum: 1
                preview_rows:
                  type: array
                  items:
                    type: object
                    description: 数据预览行
                detected_columns:
                  type: array
                  items:
                    $ref: '#/components/schemas/ColumnInfo'
                processing_time_ms:
                  type: integer
                  description: 处理耗时(毫秒)
    ColumnInfo:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum: [numerical, categorical, descriptive]
        nullable_ratio:
          type: number
          format: float
          minimum: 0
          maximum: 1
        unique_values:
          type: integer
        sample_values:
          type: array
          items:
            type: string
...
    ProcessingRequest:
      type: object
      properties:
        pca_config:
          $ref: '#/components/schemas/PCAConfig'
        tsne_config:
          $ref: '#/components/schemas/TSNEConfig'
        preprocessing_config:
          $ref: '#/components/schemas/PreprocessingConfig'
    PCAConfig:
      type: object
      properties:
        n_components:
          type: integer
          minimum: 2
          maximum: 100
          default: 50
        variance_retention:
          type: number
          format: float
          minimum: 0.8
          maximum: 1.0
          description: 保留方差比例，如果设置则忽略n_components
        whiten:
          type: boolean
          default: false
        random_state:
          type: integer
          default: 42
    TSNEConfig:
      type: object
      properties:
        perplexity:
          type: number
          format: float
          minimum: 5
          maximum: 50
          default: 30
        n_components:
          type: integer
          minimum: 2
          maximum: 3
          default: 2
        learning_rate:
          type: number
          format: float
          minimum: 10
          maximum: 1000
          default: 200
        n_iter:
          type: integer
          minimum: 250
          maximum: 5000
          default: 1000
        random_state:
          type: integer
          default: 42
        metric:
          type: string
          enum: [euclidean, manhattan, chebyshev, minkowski]
          default: euclidean
    PreprocessingConfig:
      type: object
      properties:
        missing_value_strategy:
          type: string
          enum: [drop, mean, median, mode, interpolate]
          default: mean
        scaling_method:
          type: string
          enum: [standard, minmax, robust, none]
          default: standard
        outlier_detection:
          type: boolean
          default: true
        outlier_threshold:
          type: number
          format: float
          minimum: 1.5
          maximum: 5.0
          default: 3.0
    ProcessingResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                pipeline_id:
                  type: string
                  format: uuid
                dataset_id:
                  type: string
                  format: uuid
                status:
                  type: string
                  enum: [pending, running, completed, failed]
                estimated_duration_ms:
                  type: integer
                start_time:
                  type: string
                  format: date-time
...
  responses:
    BadRequest:
      description: 请求参数错误
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ErrorResponse'
              - type: object
                properties:
                  error:
                    properties:
                      code:
                        example: INVALID_REQUEST
                      message:
                        example: 请求参数格式错误或缺少必要参数
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ErrorResponse'
              - type: object
                properties:
                  error:
                    properties:
                      code:
                        example: NOT_FOUND
                      message:
                        example: 请求的资源不存在
    FileTooLarge:
      description: 文件过大
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ErrorResponse'
              - type: object
                properties:
                  error:
                    properties:
                      code:
                        example: FILE_TOO_LARGE
                      message:
                        example: 上传的文件超过最大限制(100MB)
    UnsupportedMediaType:
      description: 不支持的媒体类型
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ErrorResponse'
              - type: object
                properties:
                  error:
                    properties:
                      code:
                        example: UNSUPPORTED_MEDIA_TYPE
                      message:
                        example: 仅支持CSV格式文件
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - BearerAuth: []
...